+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- [Added] Allow multiple instances of Crawler sharing the same queue
 - [Improved] Logger will now log entries as `debug` or `warn`
 
 ## v1.3.0 [2023-09-30]